Controllo generale, sistemati layout e css

+++ GestioneCorsi/src/it/betacom/businesscomponent/model/Docente.java
@@ -1,35 +1,40 @@
 package it.betacom.businesscomponent.model;
 
 public class Docente {
-	private int cod_docente;
+	private long cod_docente;
 	private String nomeDocente;
 	private String cognomeDocente;
 	private String cvDocente;
-	
-	public int getCod_docente() {
+
+	public long getCod_docente() {
 		return cod_docente;
 	}
-	public void setCod_docente(int cod_docente) {
+
+	public void setCod_docente(long cod_docente) {
 		this.cod_docente = cod_docente;
 	}
+
 	public String getNomeDocente() {
 		return nomeDocente;
 	}
+
 	public void setNomeDocente(String nomeDocente) {
 		this.nomeDocente = nomeDocente;
 	}
+
 	public String getCognomeDocente() {
 		return cognomeDocente;
 	}
+
 	public void setCognomeDocente(String cognomeDocente) {
 		this.cognomeDocente = cognomeDocente;
 	}
+
 	public String getCvDocente() {
 		return cvDocente;
 	}
+
 	public void setCvDocente(String cvDocente) {
 		this.cvDocente = cvDocente;
 	}
-	
-	
 }
